scaffold

+++ PitchBooking/PitchBooking.Data/Models/Pitch.cs
@@ -20,7 +20,7 @@ namespace PitchBooking.Data.Models
         public string ImgPath { get; set; }
         public TimeSpan? TimeStart { get; set; }
         public TimeSpan? TimeEnd { get; set; }
-        public int? Rates { get; set; }
+        public double? Rates { get; set; }
         public int? Status { get; set; }
         public string Phone { get; set; }
